Group-Project-AI-Engine-Test/Group-Project-AI-Engine-Test/requirements.txt
pandas
numpy
yfinance
scikit-learn
tensorflow
matplotlib
flask
flask-cors
langchain-groq
langchain-ollama
langchain-openai
crewai
pydantic
firebase-admin
python-dotenv
PyYAML
opentelemetry-api
opentelemetry-sdk
opentelemetry-proto
